# [CPDNPQ-2395] improve participants API load times