--- a/config/initializers/pagy.rb
+++ b/config/initializers/pagy.rb
@@ -1,3 +1,4 @@
+require "pagy/extras/countless"
 require "pagy/extras/overflow"
 
 Pagy::DEFAULT[:size] = [1, 1, 1, 1]