### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-2395

requests to `/api/v3/participants/npq?page_size=3000&updated_since=2020-01-01` are too slow (~800ms)

### Changes proposed in this pull request

* switch to [pagy_countless](https://ddnexus.github.io/pagy/docs/extras/countless/) (saves 200-300ms)
* add index to applications lead_provider_approval_status (saves ~100ms)
* add index to users created_at (saves ~200ms)

Tested using a production data dump, using lead provider _Ambition Institute_, which has ~46,000 users.

before optimisation: page load ~800ms
after optimisation: page load ~150ms